Move some symbols around, add typing shortcuts
Apparently I use the pipe character all the time but it's way too far
away on a standard keyboard for my small little hands so I moved it to
the raise layer on the P key.  I also moved the backslash key to the
raise layer on the forward slash key.

Also added two new macros on the comma and period keys that on the
raise layer produce "<-" and "->" respectively.