Move caps lock back to the correct location
Turns out, with shift on the home row it's really hard to type
multiple capitals in a row.  In order to keep a good flow of typing
the caps lock needs to be easy to get to.  So, restore the location of
the caps lock and move the grave key down one.